Throw exception when encountering bad mnemonic
Also made a number of changes to improve error logging and debugging. E.g.  previously labels where recorded quite haphazardly. In this updated version labels are only recorded if they are actually used.

Why is that? Because that makes it easier to determine if an unreckognized word is an opcode or label, so we can give better error reporting.